% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_prediction_quantiles}
\alias{compute_prediction_quantiles}
\title{Compute Average Predictions by Quantile}
\usage{
compute_prediction_quantiles(predictions, predicted_col, actual_col)
}
\arguments{
\item{predictions}{Data frame of predictions.}

\item{predicted_col}{Column name of predicted values.}

\item{actual_col}{Column name of actual values.}
}
\description{
Compute the average predicted and actual values by decile
}
