% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_actual_vs_expected}
\alias{plot_actual_vs_expected}
\title{Plot Actual vs. Predicted}
\usage{
plot_actual_vs_expected(df, orientation = "portrait")
}
\arguments{
\item{df}{Data frame of predictions and actuals by quantile. Should be the
return value of `compute_prediction_quantiles()`.}

\item{orientation}{this specify the orientation of the graph, Portrait or landscape.
Default value is Portrait}
}
\description{
Plot Actual vs. Predicted
}
