% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weighted_rmse}
\alias{weighted_rmse}
\title{Weighted RMSE}
\usage{
weighted_rmse(data, truth, estimate, weights)
}
\arguments{
\item{data}{The data frame with predicted and actual values.}

\item{truth}{Column name for actual values.}

\item{estimate}{Column name for predicted values.}

\item{weights}{Column name for weights.}
}
\description{
Calculates weighted RMSE
}
